Removed confirmation page
